--- src/main/java/com/social_media/social_media/controller/FollowingResponseDto.java
@@ -1,5 +1,0 @@
-package com.social_media.social_media.controller;
-
-public class FollowingResponseDto {
-
-}
