--- src/main/java/com/social_media/social_media/dto/request/PostProductRequestDto.java
@@ -1,20 +1,0 @@
-package com.social_media.social_media.dto.request;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDate;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-public class PostProductRequestDto {
-    private Long user_id;
-    @JsonFormat(pattern = "dd-MM-yyyy")
-    private LocalDate date;
-    private ProductRequestDto product;
-    private Integer category;
-    private Double price;
-}
