--- src/main/java/com/social_media/social_media/dto/request/ProductRequestDto.java
@@ -1,19 +1,0 @@
-package com.social_media.social_media.dto.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class ProductRequestDto {
-    private Long product_id;
-    private String product_name;
-    private String type;
-    private String brand;
-    private String color;
-    private String notes;
-}
