+++ src/main/java/com/social_media/social_media/dto/responseDto/FollowedResponseDto.java
@@ -2,6 +2,8 @@ package com.social_media.social_media.dto.responseDto;
 
 import java.util.List;
 
+import com.social_media.social_media.entity.User;
+
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -12,5 +14,5 @@ import lombok.NoArgsConstructor;
 public class FollowedResponseDto {
     private Long user_id;
     private String user_name;
-    private List<UserResponseDto> followed;
+    private List<User> followed;
 }
