+++ src/main/java/com/social_media/social_media/dto/responseDto/FollowersCountResponseDto.java
@@ -1,14 +1,12 @@
 package com.social_media.social_media.dto.responseDto;
 
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-@Builder
 public class FollowersCountResponseDto {
     private Long user_id;
     private String user_name;
