+++ src/main/java/com/social_media/social_media/dto/responseDto/PostResponseDto.java
@@ -1,22 +1,17 @@
 package com.social_media.social_media.dto.responseDto;
 
-import java.time.LocalDate;
+import java.util.Date;
 
-import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-@Builder
 public class PostResponseDto {
-    private Long post_id;
     private Long user_id;
-    @JsonFormat(pattern = "dd-MM-yyyy")
-    private LocalDate date;
+    private Date date;
     private ProductResponseDto product;
     private Integer category;
     private Double price;
