--- src/main/java/com/social_media/social_media/dto/responseDto/PostResponseWithIdDto.java
@@ -1,23 +1,0 @@
-package com.social_media.social_media.dto.responseDto;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDate;
-
-@Builder
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-public class PostResponseWithIdDto {
-    private Long user_id;
-    private Long post_id;
-    @JsonFormat(pattern = "dd-MM-yyyy")
-    private LocalDate date;
-    private ProductResponseDto product;
-    private Integer category;
-    private Double price;
-}
