+++ src/main/java/com/social_media/social_media/dto/responseDto/ProductResponseDto.java
@@ -1,11 +1,9 @@
 package com.social_media.social_media.dto.responseDto;
 
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@Builder
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
