+++ src/main/java/com/social_media/social_media/entity/Follow.java
@@ -1,14 +1,12 @@
 package com.social_media.social_media.entity;
 
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-@Builder
 public class Follow {
     private Long followerId;
     private Long followedId;
