+++ src/main/java/com/social_media/social_media/entity/Post.java
@@ -1,23 +1,19 @@
 package com.social_media.social_media.entity;
 
-import java.time.LocalDate;
+import java.util.Date;
 
-import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-@Builder
 public class Post {
-    private Long postId;
     private Long userId;
-    @JsonFormat(pattern = "dd-MM-yyyy")
-    private LocalDate date;
+    private Date date;
     private Product product;
     private Integer category;
     private Double price;
+
 }
