+++ src/main/java/com/social_media/social_media/entity/Product.java
@@ -1,14 +1,12 @@
 package com.social_media.social_media.entity;
 
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-@Builder
 public class Product {
     private Long productId;
     private String productName;
