+++ src/main/java/com/social_media/social_media/entity/User.java
@@ -1,12 +1,8 @@
 package com.social_media.social_media.entity;
 
-import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 
 @Data
-@AllArgsConstructor
-@NoArgsConstructor
 public class User {
     private Long userId;
     private String name;
