+++ src/main/java/com/social_media/social_media/exception/ExceptionController.java
@@ -9,21 +9,9 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 @ControllerAdvice
 public class ExceptionController {
     @ExceptionHandler(NotFoundException.class)
-    public ResponseEntity<ExceptionDto> notFound(NotFoundException e) {
+    public ResponseEntity<?> notFound(NotFoundException e) {
         ExceptionDto exceptionDto = new ExceptionDto(e.getMessage());
         return new ResponseEntity<>(exceptionDto, HttpStatus.NOT_FOUND);
-    }
-
-    @ExceptionHandler(BadRequestFollowException.class)
-    public ResponseEntity<ExceptionDto> badRequest(BadRequestFollowException e) {
-        ExceptionDto exceptionDto = new ExceptionDto(e.getMessage());
-        return new ResponseEntity<>(exceptionDto, HttpStatus.BAD_REQUEST);
-    }
-
-    @ExceptionHandler(NotSellerException.class)
-    public ResponseEntity<?> notSeller(NotSellerException e) {
-        ExceptionDto exceptionDto = new ExceptionDto(e.getMessage());
-        return new ResponseEntity<>(exceptionDto, HttpStatus.FORBIDDEN);
     }
 
 }
