--- src/main/java/com/social_media/social_media/exception/NotSellerException.java
@@ -1,7 +1,0 @@
-package com.social_media.social_media.exception;
-
-public class NotSellerException extends RuntimeException {
-    public NotSellerException(String message) {
-        super(message);
-    }
-}
