+++ src/main/java/com/social_media/social_media/repository/user/UserRepositoryImpl.java
@@ -2,9 +2,7 @@ package com.social_media.social_media.repository.user;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.social_media.social_media.dto.responseDto.UserResponseDto;
 import com.social_media.social_media.entity.User;
-import com.social_media.social_media.exception.NotFoundException;
 import org.springframework.stereotype.Repository;
 import org.springframework.util.ResourceUtils;
 
@@ -12,7 +10,6 @@ import java.io.File;
 import java.io.IOException;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Repository
@@ -39,10 +36,5 @@ public class UserRepositoryImpl implements IUserRepository {
         });
 
         users = userList.stream().collect(Collectors.toMap(User::getUserId, user -> user));
-    }
-
-    @Override
-    public Optional<User> findById(Long userId) {
-        return Optional.ofNullable(this.users.get(userId));
     }
 }
