+++ src/main/resources/follows.json
@@ -1,6 +1,5 @@
 [
   { "followerId": 2, "followedId": 1 },
-  { "followerId": 2, "followedId": 3 },
   { "followerId": 3, "followedId": 1 },
   { "followerId": 4, "followedId": 2 },
   { "followerId": 5, "followedId": 2 },
@@ -8,8 +7,6 @@
   { "followerId": 7, "followedId": 3 },
   { "followerId": 8, "followedId": 4 },
   { "followerId": 9, "followedId": 4 },
-  { "followerId": 1, "followedId": 6 },
-  { "followerId": 1, "followedId": 7 },
-  { "followerId": 1, "followedId": 8 },
-  { "followerId": 1, "followedId": 9 }
+  { "followerId": 10, "followedId": 5 },
+  { "followerId": 1, "followedId": 6 }
 ]
